// src/app/services/colaboradores/model/colaboradores.model.ts
export interface ColaboradoresModel {
  colaboradores: ColaboradorModel[];
}

export interface ColaboradorModel {
  nomfun: string;
  nomusu: string;
  razsoc: string;
  nomloc: string;
  titcar: string;
  nomfil: string;
  destip: string;
  resccu: string;
  numcad: string;
  filSap: string;
  empSap: string;
  tipcol: string;
}